refactor: class-pg-notifications away as wp_queue handles thta
